Consolidates logic for firebase ops within the useTasks composable

+++ src/composables/useTasks.ts
@@ -1,6 +1,6 @@
 import { ref, type Ref } from 'vue';
 import { useCollection } from 'vuefire';
-import { deleteDoc, updateDoc, FirestoreError } from '@firebase/firestore';
+import { deleteDoc, updateDoc, addDoc, FirestoreError } from '@firebase/firestore';
 import { taskDocRef, taskListRef } from '@/api';
 
 import Task from '@/models/Task';
@@ -13,9 +13,9 @@ const isLoading = ref<boolean>(false);
 type UseTasks = {
   tasks: Ref<Task[]>;
   isLoading: Ref<boolean>;
-  fetchTasks: () => Promise<void>;
-  nextStatus: (id: Task['id']) => Promise<void>;
-  deleteTask: (id: Task['id']) => Promise<void>;
+  fetchTasks: () => Promise<any>;
+  nextStatus: (id: Task['id']) => Promise<any>;
+  deleteTask: (id: Task['id']) => Promise<any>;
 };
 
 export default function (): UseTasks {
@@ -35,17 +35,33 @@ export default function (): UseTasks {
     }
   };
 
-  const deleteTask = async (id: Task['id']) => {
+  const updateTask = async (id?: Task['id'], task?: Task) => {
+    const isTaskEmpty = task && Object.keys(task).length > 0;
     isLoading.value = true;
 
     try {
       error.value = '';
-      return await deleteDoc(taskDocRef(id));
+
+      if (id && !isTaskEmpty) {
+        // Firebase's web client library doesn't support objects instantiated with new...()
+        // Hence the explicit cloning as as POJO. https://stackoverflow.com/a/48158848
+        return await updateDoc(taskDocRef(id), { ...task });
+      } else if (!id && !isTaskEmpty) {
+        return await addDoc(taskListRef, { ...task });
+      } else if (id && isTaskEmpty) {
+        return await deleteDoc(taskDocRef(id));
+      } else if (!id && isTaskEmpty) {
+        throw new Error('No ID or task data provided while updating task list');
+      }
     } catch (exception) {
       error.value = exception as string | FirestoreError | Error;
     } finally {
       isLoading.value = false;
     }
+  };
+
+  const deleteTask = async (id: Task['id']) => {
+    return await updateTask(id);
   };
 
   const nextStatus = async (id: Task['id']) => {
@@ -67,19 +83,7 @@ export default function (): UseTasks {
         task.status = nextStatus;
       }
 
-      isLoading.value = true;
-
-      try {
-        error.value = '';
-
-        // Firebase's web client library doesn't support objects instantiated with new...()
-        // Hence the explicit cloning as as POJO. https://stackoverflow.com/a/48158848
-        return await updateDoc(taskDocRef(id), { ...task });
-      } catch (exception) {
-        error.value = exception as string | FirestoreError | Error;
-      } finally {
-        isLoading.value = false;
-      }
+      return updateTask(id, task);
     }
   };
